Add files via upload
I have tried to create face detection with rectangular green to detect face by using python code. as I have run it in vs code and to stop the live output you need to click "a" and it automatically stops. I have referred some videos and tried to build in my laptop.